src/main/java/com/iryna/db/mapper/GenreRowMapper.java
package com.iryna.db.mapper;

import com.iryna.entity.Genre;
import org.springframework.jdbc.core.RowMapper;

import java.sql.ResultSet;
import java.sql.SQLException;

public class GenreRowMapper implements RowMapper<Genre> {
    @Override
    public Genre mapRow(ResultSet resultSet, int i) throws SQLException {
        long id = resultSet.getLong("id");
        String name = resultSet.getString("name");

        return Genre
                .builder()
                .id(id)
                .name(name)
                .build();
    }
}
